+++ lib/tentd/streaming/puma_fix.rb
@@ -1,0 +1,12 @@
+# Fix Puma treating empty BodyProxies as non-empty
+module Rack
+  class BodyProxy
+    def ==(other)
+      if other == [] and @body == []
+        true
+      else
+        super
+      end
+    end
+  end
+end
